Move helpers to seperate file, and reporting into seperate module
